requirements.txt
# application requirements
djangorestframework==3.12.4
django-environ==0.4.5
django-extensions
django-filter==2.4.0
markdown==3.3.4
psycopg2==2.9.1
gunicorn==20.1.0
gevent==21.1.2
jsonschema==3.2.0
drf-yasg==1.20.0
django-storages==1.11.1
django==3.2.5
pandas
ipython

# development requirements
